[python] extract converter_dunder.cpp from python_converter.cpp
Move 7 dunder-dispatch methods of python_converter (op_to_dunder,
op_to_rdunder, find_dunder_method, has_dunder_method, build_dunder_call,
dispatch_dunder_operator, dispatch_unary_dunder_operator) along with
their three file-local static helpers (is_excluded_struct_tag,
resolve_operand_type, is_other_param_compatible) into
src/python-frontend/converter/converter_dunder.cpp.

Pure code movement — no logic change.
