+++ src/python-frontend/CMakeLists.txt
@@ -15,6 +15,7 @@ add_library(pythonfrontend STATIC
             python_converter.cpp
             converter/converter_class.cpp
             converter/converter_compare.cpp
+            converter/converter_dunder.cpp
             converter/converter_symbols.cpp
             converter/converter_types.cpp
             converter/converter_util.cpp
